requirements.txt
# CaseMind - Legal Case Similarity Search System
# Python 3.9+ Required

# ========================================================================
# Core AI/ML Libraries
# ========================================================================
openai>=1.0.0                          # LLM API for metadata/fact extraction
sentence-transformers>=2.2.0           # Embedding models
transformers>=4.30.0                   # Transformer models and cross-encoders
torch>=2.0.0                           # PyTorch backend
numpy>=1.24.0                          # Numerical computing
scikit-learn>=1.3.0                    # ML utilities

# ========================================================================
# Haystack Framework
# ========================================================================
haystack-ai>=2.0.0                     # Pipeline framework
farm-haystack[inference]>=1.22.0       # Haystack with inference components

# ========================================================================
# Database and Storage
# ========================================================================
psycopg2-binary>=2.9.9                 # PostgreSQL adapter
pgvector>=0.2.0                        # PostgreSQL vector extension Python client

# ========================================================================
# PDF Processing
# ========================================================================
PyMuPDF>=1.23.0                        # PDF text extraction (fitz)

# ========================================================================
# Configuration and Environment
# ========================================================================
python-dotenv>=1.0.0                   # Environment variable management

# ========================================================================
# CLI and UI
# ========================================================================
rich>=13.0.0                           # Rich terminal UI

# ========================================================================
# Utilities
# ========================================================================
pathlib>=1.0.1                         # Path handling (usually built-in)

# ========================================================================
# Development and Testing (Optional)
# ========================================================================
pytest>=7.0.0                          # Testing framework
pytest-cov>=4.1.0                      # Coverage reporting
pytest-asyncio>=0.21.0                 # Async test support
black>=23.0.0                          # Code formatting
flake8>=6.0.0                          # Linting
mypy>=1.0.0                            # Type checking

